+++ src/Resources/Guests/Baskets.ts
@@ -1,12 +1,5 @@
 import AbstractResource from '../AbstractResource';
 
-/**
- * _Used for **guest** baskets only._
- * - Getting a guest basket: `client.guests.baskets.get(basketId);`
- * - Adding a basketLine: `client.guests.baskets.addLine({ basketId, productVariantId, quantity });`
- * - Updating a basketLine: `client.guests.baskets.updateLine({ basketLineId, quantity });`
- * - Deleting a basketLine `client.guests.baskets.removeLine(basketLineId);`
- */
 export default class Baskets extends AbstractResource {
     initialise(): void {
         this.resourcePath = '/basket-lines';
