+++ src/Resources/Users/Baskets.ts
@@ -1,12 +1,5 @@
 import AbstractResource from '../AbstractResource';
 
-/**
- * _Used for **member's** baskets only._
- * - Getting a user's basket: `client.user.basket.getActive();`
- * - Adding a basketLine: `client.user.basket.addLine({ productVariantId, quantity });`
- * - Updating a basketLine: `client.user.basket.updateLine({ basketLineId, quantity });`
- * - Deleting a basketLine: `client.user.basket.removeLine(basketLineId);`
- */
 export default class Baskets extends AbstractResource {
     initialise(): void {
         this.resourcePath = '';
