+++ src/Resources/Users/Details.ts
@@ -1,10 +1,5 @@
 import AbstractResource from '../AbstractResource';
 
-/**
- * _Updating a user's details_
- *
- * `client.user.details.update();`
- */
 export default class Details extends AbstractResource {
     initialise(): void {
         this.resourcePath = '/users';
