+++ src/Resources/Users/Toybox.ts
@@ -1,11 +1,5 @@
 import AbstractResource from '../AbstractResource';
 
-/**
- * _Updating a user's toybox_
- *
- * - `client.user.toybox.designateToyForReturn(stockId);`
- * - `client.user.toybox.undesignateToyForReturn(stockId);`
- */
 export default class Toybox extends AbstractResource {
     initialise(): void {
         this.resourcePath = '/stock';
